scripts: Move non-common functions to tests.py
